Support deauthentication codes (#349)
* Implement deauth reason and assoc status

* Log unhandled request when only get is implemented but set is called or vice versa

* Add noise info in scan results

* Fill known capabilities based on ic_caps
Order and squash other capability bytes
Don't set bytes 8-9
bytes 0, 1 and 7 alone are sufficient for Apple Watch unlock. It can only unlock administrative permissions and doesn't try to unlock the lock screen, because AWDL capability is disabled (something in bytes 2-6). If these bytes are set, then it tries to unlock the lock screen too but fails, because AWDL features are not implemented.

Co-authored-by: Michael Belyaev <usrsse2@me.com>